requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
defusedxml==0.7.1
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-xml==2.0.0
Pillow==9.3.0
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
